+++ lib/src/commands/update_command.dart
@@ -6,9 +6,7 @@ import 'package:maco/src/version.dart';
 import 'package:mason_logger/mason_logger.dart';
 import 'package:pub_updater/pub_updater.dart';
 
-/// {@template update_command}
 /// A command which updates the CLI.
-/// {@endtemplate}
 class UpdateCommand extends Command<int> {
   /// {@macro update_command}
   UpdateCommand({