+++ lib/src/messages.dart
@@ -1,0 +1,21 @@
+Map<String, String> en = {
+  'beg_ext': 'Begging video extraction.',
+  'using_def_token': 'Token not set, using default.',
+  'non_null_url': 'Url option cannot be null.',
+  'error_occ': 'An error occurred',
+  'downloading': 'Downloading.',
+  'error_dl': 'An error occurred while downloading',
+  "error_log": 'Error:',
+  'dl_at': 'Downloaded at'
+};
+
+Map<String, String> ru = {
+  'beg_ext': 'Начинаю вытаскивание ссылки.',
+  'using_def_token': 'Токен не указан, изпользую стандартный.',
+  'non_null_url': 'Поле url обязательно.',
+  'error_occ': 'Произошла ошибка',
+  'downloading': 'Скачиваю.',
+  'error_dl': 'Произошла ошибка при скачивании.',
+  "error_log": 'Ошибка:',
+  'dl_at': 'Скачано в'
+};