+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.2
+
+* Sibnet parser!
+* Source code optimizations
+
 ## 0.0.1
 
 * Release!